refactor: modularize sections and add routing
